♻️ address review feedback on main-communication language restriction
- Fix correctness gap: a legacy/out-of-set language in mainCommunication
  (saved before this restriction existed) was silently dropped from the
  resolved title set, so an unresolvable selection incorrectly read as
  "none selected" and passed validation. Now flagged as invalid instead.
- De-duplicate the numeric-id-vs-translated-title language resolution
  logic, previously copy-pasted three times (NewOpportunity.tsx,
  OpportunityDetailsEdit.tsx, opportunityDetailsSchema.ts), into a single
  resolveFormLanguageToOption helper in formatters.ts.
- Prefer isoCode ("de"/"en") over a hardcoded title string match in
  getMainCommunicationLanguageOptions when the option list provides it.
- Reword the validation message for clarity ("German only" vs "German").

Addresses review feedback on need4deed-org/fe#829.
