+++ public/locales/en/translations.json
@@ -1295,7 +1295,7 @@
           "descriptionTooLong": "Description is too long",
           "numberOfVolunteersRequired": "Number of volunteers must be at least 1",
           "languageRequired": "At least one language is required",
-          "mainCommunicationInvalid": "Main communication must be German, or German and English",
+          "mainCommunicationInvalid": "Main communication must be German only, or German and English",
           "availabilityRequired": "At least one time slot must be selected",
           "activitiesRequired": "At least one activity is required",
           "skillsRequired": "At least one skill is required"
