+++ src/components/Dashboard/NewOpportunity/NewOpportunity.tsx
@@ -14,6 +14,7 @@ import {
   getMainCommunicationLanguageOptions,
   NewOpportunityDetailsFormData,
 } from "@/components/Dashboard/Profile/sections/OpportunityDetails/opportunityDetailsSchema";
+import { resolveFormLanguageToOption } from "@/components/Dashboard/Profile/sections/OpportunityDetails/formatters";
 import { AccompanyingDetailsEdit } from "@/components/Dashboard/Profile/sections/AccompanyingDetails/AccompanyingDetailsEdit";
 import { FormDetails } from "@/components/Dashboard/Profile/sections/shared/styles";
 import { BackButton, PageContainer } from "@/components/Dashboard/Profile/styles";
@@ -107,18 +108,7 @@ function toLangOptionItems(
   t: TFunction,
 ): OptionItem[] {
   return formLangs.flatMap(({ language }) => {
-    if (!language) return [];
-    const numId = Number(language);
-    if (!isNaN(numId) && numId > 0) {
-      const found = apiLanguages.find((a) => a.id === numId);
-      return found ? [{ id: found.id, title: found.title }] : [];
-    }
-    const found = apiLanguages.find((a) => {
-      if (a.title === language || a.title.toLowerCase() === language.toLowerCase()) return true;
-      const key = `languageNames.${a.title.toLowerCase()}`;
-      const translated = t(key);
-      return translated !== key && translated === language;
-    });
+    const found = resolveFormLanguageToOption(language, apiLanguages, t);
     return found ? [{ id: found.id, title: found.title }] : [];
   });
 }
